+++ src/onetick/py/__init__.py
@@ -192,7 +192,7 @@ from onetick.py.sources import (Tick, TTicks, Ticks, Orders, Trades, NBBO, Quote
                                 Custom, query, Symbols, Empty, DataSource, LocalCSVTicks, SymbologyMapping,
                                 ObSnapshot, ObSnapshotWide, ObSnapshotFlat, ObSummary, ObSize, ObVwap, ObNumLevels,
                                 by_symbol, ODBC, SplitQueryOutputBySymbol, DataFile, PointInTime,
-                                ReadSnapshot, ShowSnapshotList, FindSnapshotSymbols)
+                                ReadSnapshot, ShowSnapshotList, FindSnapshotSymbols, ReadFromDataFrame)
 from onetick.py.utils import adaptive, range, perf
 from onetick.py.session import Session, TestSession, Config, Locator, HTTPSession
 from onetick.py.servers import RemoteTS, LoadBalancing, FaultTolerance